Добавил хендлер для NoResultFound.

+++ src/routers/menu.py
@@ -70,10 +70,10 @@ async def get_menu(
 ):
     """Эндпойнт для получения определенного меню."""
 
-    try:
-        menu = await get_menu_db(target_menu_id, session)
-    except NoResultFound:
-        return JSONResponse(content={'detail': 'menu not found'}, status_code=HTTP_404_NOT_FOUND)
+    menu = await get_menu_db(target_menu_id, session)
+    if not menu:
+        NoResultFound.args = 'menu'
+        raise NoResultFound
     return menu
 
 
@@ -92,10 +92,10 @@ async def update_menu(
 
     query = select(Menu).where(Menu.id == target_menu_id)
     result = await session.execute(query)
-    try:
-        menu = result.scalars().unique().one()
-    except NoResultFound:
-        return JSONResponse(content={'detail': 'menu not found'}, status_code=HTTP_404_NOT_FOUND)
+    menu = result.scalars().unique().one_or_none()
+    if not menu:
+        NoResultFound.args = 'menu'
+        raise NoResultFound
     menu.title = data.title
     menu.description = data.description
     await session.commit()
@@ -118,10 +118,10 @@ async def delete_menu(
 
     query = select(Menu).where(Menu.id == target_menu_id)
     result = await session.execute(query)
-    try:
-        menu = result.scalars().unique().one()
-    except NoResultFound:
-        return JSONResponse(content={'detail': 'menu not found'}, status_code=HTTP_404_NOT_FOUND)
+    menu = result.scalars().unique().one_or_none()
+    if not menu:
+        NoResultFound.args = 'menu'
+        raise NoResultFound
     await session.delete(menu)
     await session.commit()
     return JSONResponse(content={'message': 'Success.'}, status_code=HTTP_200_OK)
